store subscriptions in the LCM, and add a 3-argument subscribe
We need to make sure the subscription callbacks stay in memory, so we
hold a reference to them inside the LCM object.

Also, adds a 3-argument subscribe() method which does not take a message
type and just calls the handler with the raw bytes of the message. This
makes PyLCM really easy to implement